refactoring of resolvers for modularity, helper function created to deal with date property of data
