+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 nodemon.json
 package-lock.json
+.vscode
+.DS_Store
